pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Analises web pages"
authors = ["artem <artem_smirnov88@mail.ru>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10.12"
flask = "^2.2.3"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.1"
validators = "^0.34.0"
psycopg2-binary ="^2.9.10"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
